Hard-code the original Cobra help template
Cobra 1.6 has modified the default help template to support grouping.
This breaks the logic in cobrautils which manipulates that help string.

This commit hard-codes the original cobra help string instead of
obtaining it from Cobra.  This allows cobrautils to continue working as
before for older versions of Cobra or newer ones.

Signed-off-by: Marc Khouzam <marc.khouzam@gmail.com>